requirements.txt
ffmpeg-python
requests
python-dotenv
flask
flask_cors
opencv-python
numpy
pillow
rembg

